chore: Move mjpeg-server import to dev dependencies
