+++ package.json
@@ -44,7 +44,6 @@
     "klaw": "^3.0.0",
     "lockfile": "^1.0.4",
     "lodash": "^4.2.1",
-    "mjpeg-server": "^0.3.0",
     "mkdirp": "^1.0.0",
     "moment": "^2.24.0",
     "mv": "^2.1.1",
@@ -88,6 +87,7 @@
     "chai-as-promised": "^7.1.1",
     "eslint-config-appium": "^4.2.0",
     "gulp": "^4.0.0",
+    "mjpeg-server": "^0.3.0",
     "mocha": "^8.0.1",
     "pre-commit": "^1.1.3",
     "sinon": "^10.0.0"
